+++ docs/src/multiplayer/player1.js
@@ -43,6 +43,7 @@ export default {
     game: {
       type: "game",
       multiplayer: {
+        // serverUrl: "ws://localhost:3000",
         serverUrl: "wss://inglorious-server.onrender.com",
       },
       devMode: true,