+++ examples/apps/todomvc-cs/src/store/index.js
@@ -5,12 +5,14 @@ import { entities } from "./entities"
 import { types } from "./types"
 import { createReactStore } from "@inglorious/react-store"
 
-export const store = createStore({ types, entities })
+const config = { mode: "batched", skippedEvents: ["create"] }
 
-export const { Provider, useSelector, useNotify } = createReactStore(store, {
-  mode: "batched",
-  skippedEvents: ["create"],
-})
+export const store = createStore({ ...config, types, entities })
+
+export const { Provider, useSelector, useNotify } = createReactStore(
+  store,
+  config,
+)
 
 if (import.meta.env.DEV) {
   connectDevTools(store)