+++ packages/react-store/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @inglorious/react-store
 
+## 3.0.0
+
+### Minor Changes
+
+- Add eager mode by default on store
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@5.4.0
+
+## 2.0.0
+
+### Minor Changes
+
+- Enable eager mode by default on store
+
+### Patch Changes
+
+- Updated dependencies
+  - @inglorious/store@5.3.0
+
 ## 1.0.2
 
 ### Patch Changes